Fixed bug where variables on the right caused an error
If the variable was ever on the right side of the equation the program would quit. This commit changes that by switching the perspective the computer has on the equation by switching the left and right sides so the program can handle it. Now it can solve linear equations regardless of what data type each node in the graph is.